Move helpers from the metadata pkg to either the internal type, or local to the caller
