+++ cmd/ignite-spawn/ignite-spawn.go
@@ -57,16 +57,16 @@ func StartVM(co *options) error {
 
 	// VM state handling
 	// TODO: Use a .Patch here instead
-	if err := setState(api.VMStateRunning); err != nil {
+	if err := setState(co.vm.VM, api.VMStateRunning); err != nil {
 		return fmt.Errorf("failed to update VM state: %v", err)
 	}
-	defer setState(api.VMStateStopped) // Performs a save, all other metadata-modifying defers need to be after this
+	defer setState(co.vm.VM, api.VMStateStopped) // Performs a save, all other metadata-modifying defers need to be after this
 
 	// Remove the snapshot overlay post-run, which also removes the detached backing loop devices
 	defer co.vm.RemoveSnapshot()
 
 	// Remove the IP addresses post-run
-	defer co.vm.ClearIPAddresses()
+	defer clearIPAddresses(co.vm.VM)
 
 	// Remove the Prometheus socket post-run
 	defer os.Remove(metricsSocket)
@@ -83,4 +83,9 @@ func setState(vm *api.VM, s api.VMState) error {
 	vm.Status.State = s
 
 	return client.VMs().Set(vm)
+}
+
+func clearIPAddresses(vm *api.VM) {
+	vm.Status.IPAddresses = nil
+	// TODO: This currently relies on the ordering of a set of defers. Make this more robust in the future.
 }
