+++ pkg/apis/ignite/helpers.go
@@ -1,5 +1,11 @@
 package ignite
 
+import (
+	"path"
+
+	"github.com/weaveworks/ignite/pkg/constants"
+)
+
 // GetNetworkModes gets the list of available network modes
 func GetNetworkModes() []NetworkMode {
 	return []NetworkMode{
@@ -34,4 +40,33 @@ func (vm *VM) SetImage(image *Image) {
 func (vm *VM) SetKernel(kernel *Kernel) {
 	vm.Spec.Kernel.OCIClaim = kernel.Spec.OCIClaim
 	vm.Status.Kernel = kernel.Status.OCISource
+}
+
+// SnapshotDev returns the path where the (legacy) DM snapshot exists
+func (vm *VM) SnapshotDev() string {
+	// TODO: Reuse the prefixer here
+	return path.Join("/dev/mapper", constants.IGNITE_PREFIX+vm.GetUID().String())
+}
+
+// Running returns true if the VM is running, otherwise false
+func (vm *VM) Running() bool {
+	return vm.Status.State == VMStateRunning
+}
+
+// ObjectPath returns the directory where this VM's data is stored
+func (vm *VM) ObjectPath() string {
+	// TODO: Move this into storage
+	return path.Join(constants.DATA_DIR, vm.GetKind().Lower(), vm.GetUID().String())
+}
+
+// ObjectPath returns the directory where this Image's data is stored
+func (img *Image) ObjectPath() string {
+	// TODO: Move this into storage
+	return path.Join(constants.DATA_DIR, img.GetKind().Lower(), img.GetUID().String())
+}
+
+// ObjectPath returns the directory where this Kernel's data is stored
+func (k *Kernel) ObjectPath() string {
+	// TODO: Move this into storage
+	return path.Join(constants.DATA_DIR, k.GetKind().Lower(), k.GetUID().String())
 }
